@@ -1,8 +1,8 @@
 module pdhcp
 
-go 1.15
+go 1.20
 
 require (
-	github.com/pyke369/golang-support v0.0.0-20200321184731-cd1ed731523d
-	golang.org/x/sys v0.0.0-20200909081042-eff7692f9009
+	github.com/pyke369/golang-support v0.0.0-20230108182845-9fe2fa61d8f6
+	golang.org/x/sys v0.4.0
 )
